+++ internal/config/config_test.go
@@ -244,6 +244,20 @@ func TestLoad(t *testing.T) {
 			},
 		},
 		{
+			name: "environment variable substitution",
+			path: "./testdata/envsubst.yml",
+			envOverrides: map[string]string{
+				"HTTP_PORT":  "18080",
+				"LOG_FORMAT": "json",
+			},
+			expected: func() *Config {
+				cfg := Default()
+				cfg.Log.Encoding = "json"
+				cfg.Server.HTTPPort = 18080
+				return cfg
+			},
+		},
+		{
 			name: "deprecated tracing jaeger",
 			path: "./testdata/deprecated/tracing_jaeger.yml",
 			expected: func() *Config {
@@ -649,12 +663,12 @@ func TestLoad(t *testing.T) {
 
 				cfg.Authorization = AuthorizationConfig{
 					Required: true,
-					Policy: &AuthorizationSourceConfig{
-						Backend: AuthorizationBackendLocal,
-						Local: &AuthorizationLocalConfig{
-							Path: "/path/to/policy.rego",
+					Backend:  AuthorizationBackendLocal,
+					Local: &AuthorizationLocalConfig{
+						Policy: &AuthorizationSourceLocalConfig{
+							Path:         "/path/to/policy.rego",
+							PollInterval: 5 * time.Minute,
 						},
-						PollInterval: 30 * time.Second,
 					},
 				}
 
@@ -721,7 +735,6 @@ func TestLoad(t *testing.T) {
 					},
 				}
 
-				cfg.Experimental.Authorization.Enabled = true
 				return cfg
 			},
 		},
@@ -802,7 +815,7 @@ func TestLoad(t *testing.T) {
 
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendMemory,
 						},
@@ -897,23 +910,19 @@ func TestLoad(t *testing.T) {
 
 				cfg.Authorization = AuthorizationConfig{
 					Required: true,
-					Policy: &AuthorizationSourceConfig{
-						Backend: AuthorizationBackendLocal,
-						Local: &AuthorizationLocalConfig{
-							Path: "/path/to/policy.rego",
+					Backend:  AuthorizationBackendLocal,
+					Local: &AuthorizationLocalConfig{
+						Policy: &AuthorizationSourceLocalConfig{
+							Path:         "/path/to/policy.rego",
+							PollInterval: time.Minute,
 						},
-						PollInterval: time.Minute,
-					},
-					Data: &AuthorizationSourceConfig{
-						Backend: AuthorizationBackendLocal,
-						Local: &AuthorizationLocalConfig{
-							Path: "/path/to/policy/data.json",
+						Data: &AuthorizationSourceLocalConfig{
+							Path:         "/path/to/policy/data.json",
+							PollInterval: time.Minute,
 						},
-						PollInterval: 30 * time.Second,
 					},
 				}
 
-				cfg.Experimental.Authorization.Enabled = true
 				return cfg
 			},
 		},
@@ -957,7 +966,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: LocalStorageType,
-					Local: &Local{
+					Local: &StorageLocalConfig{
 						Path: ".",
 					},
 				}
@@ -971,7 +980,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendMemory,
 						},
@@ -991,7 +1000,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendMemory,
 						},
@@ -1012,7 +1021,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendMemory,
 						},
@@ -1033,7 +1042,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendLocal,
 							Path: "/path/to/gitdir",
@@ -1084,7 +1093,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: GitStorageType,
-					Git: &Git{
+					Git: &StorageGitConfig{
 						Backend: GitBackend{
 							Type: GitBackendMemory,
 						},
@@ -1111,9 +1120,9 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: ObjectStorageType,
-					Object: &Object{
+					Object: &StorageObjectConfig{
 						Type: S3ObjectSubStorageType,
-						S3: &S3{
+						S3: &S3Storage{
 							Bucket:       "testbucket",
 							PollInterval: time.Minute,
 						},
@@ -1129,9 +1138,9 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: ObjectStorageType,
-					Object: &Object{
+					Object: &StorageObjectConfig{
 						Type: S3ObjectSubStorageType,
-						S3: &S3{
+						S3: &S3Storage{
 							Bucket:       "testbucket",
 							Prefix:       "prefix",
 							Region:       "region",
@@ -1149,7 +1158,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: OCIStorageType,
-					OCI: &OCI{
+					OCI: &StorageOCIConfig{
 						Repository:       "some.target/repository/abundle:latest",
 						BundlesDirectory: "/tmp/bundles",
 						Authentication: &OCIAuthentication{
@@ -1171,7 +1180,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: OCIStorageType,
-					OCI: &OCI{
+					OCI: &StorageOCIConfig{
 						Repository:       "some.target/repository/abundle:latest",
 						BundlesDirectory: "/tmp/bundles",
 						Authentication: &OCIAuthentication{
@@ -1193,7 +1202,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: OCIStorageType,
-					OCI: &OCI{
+					OCI: &StorageOCIConfig{
 						Repository:       "some.target/repository/abundle:latest",
 						BundlesDirectory: "/tmp/bundles",
 						Authentication: &OCIAuthentication{
@@ -1213,7 +1222,7 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: OCIStorageType,
-					OCI: &OCI{
+					OCI: &StorageOCIConfig{
 						Repository:       "some.target/repository/abundle:latest",
 						BundlesDirectory: "/tmp/bundles",
 						PollInterval:     5 * time.Minute,
@@ -1270,9 +1279,9 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: ObjectStorageType,
-					Object: &Object{
+					Object: &StorageObjectConfig{
 						Type: AZBlobObjectSubStorageType,
-						AZBlob: &AZBlob{
+						AZBlob: &AZBlobStorage{
 							Container:    "testdata",
 							Endpoint:     "https//devaccount.blob.core.windows.net",
 							PollInterval: 5 * time.Minute,
@@ -1294,9 +1303,9 @@ func TestLoad(t *testing.T) {
 				cfg := Default()
 				cfg.Storage = StorageConfig{
 					Type: ObjectStorageType,
-					Object: &Object{
+					Object: &StorageObjectConfig{
 						Type: GSBlobObjectSubStorageType,
-						GS: &GS{
+						GS: &GSStorage{
 							Bucket:       "testdata",
 							Prefix:       "prefix",
 							PollInterval: 5 * time.Minute,
@@ -1683,20 +1692,30 @@ func TestGetConfigFile(t *testing.T) {
 	blob.DefaultURLMux().RegisterBucket("mock", &mockURLOpener{
 		bucket: memblob.OpenBucket(nil),
 	})
-	configData := []byte("some config data")
-	ctx := context.Background()
-	b, err := blob.OpenBucket(ctx, "mock://mybucket")
+
+	var (
+		ctx        = context.Background()
+		configData = []byte("some config data")
+		b, err     = blob.OpenBucket(ctx, "mock://mybucket")
+	)
+
 	require.NoError(t, err)
 	t.Cleanup(func() { b.Close() })
+
 	w, err := b.NewWriter(ctx, "config/local.yml", nil)
 	require.NoError(t, err)
+
 	_, err = w.Write(configData)
 	require.NoError(t, err)
+
 	err = w.Close()
 	require.NoError(t, err)
+
 	t.Run("successful", func(t *testing.T) {
 		f, err := getConfigFile(ctx, "mock://mybucket/config/local.yml")
 		require.NoError(t, err)
+		t.Cleanup(func() { f.Close() })
+
 		s, err := f.Stat()
 		require.NoError(t, err)
 		require.Equal(t, "local.yml", s.Name())